+++ src/app/armory/Links.tsx
@@ -5,14 +5,13 @@ import { LoreLink } from 'app/item-popup/ItemDescription';
 import { useIsPhonePortrait } from 'app/shell/selectors';
 import { getSocketsWithStyle, isWeaponMasterworkSocket } from 'app/utils/socket-utils';
 import { DestinySocketCategoryStyle } from 'bungie-api-ts/destiny2';
-import { ItemCategoryHashes } from 'data/d2/generated-enums';
+import { ItemCategoryHashes, PlugCategoryHashes } from 'data/d2/generated-enums';
 import destinysets from 'images/destinysets.svg';
 import destinytracker from 'images/destinytracker.png';
 import logo from 'images/dimlogo.svg';
 import gunsmith from 'images/gunsmith.png';
 import lightgg from 'images/lightgg.png';
 import _ from 'lodash';
-import React from 'react';
 import { useSelector } from 'react-redux';
 import styles from './Links.m.scss';
 
@@ -27,7 +26,7 @@ const links = [
     name: 'Light.gg',
     icon: lightgg,
     link: (item: DimItem, language: string) =>
-      `https://www.light.gg/db/${language}/items/${item.hash}`,
+      `https://www.light.gg/db/${language}/items/${item.hash}${buildLightGGSockets(item)}`,
   },
   { name: 'DestinyTracker', icon: destinytracker, link: destinyDBLink },
   {
@@ -108,29 +107,72 @@ function buildSocketParam(item: DimItem): string {
 }
 
 /**
- * D2Gunsmith's socket format is: [...<first four perks, padded out if necessary, masterwork, weapon mod].join(',')
+ * Light.gg's socket format is highly similar to that of D2Gunsmith: [...<first four perks, padded out if necessary, origin trait, masterwork, weapon mod].join(',')
  */
-function buildGunsmithSockets(item: DimItem) {
-  if (item.sockets) {
-    const perkValues: number[] = [0, 0, 0, 0];
-    const perks = getSocketsWithStyle(item.sockets, DestinySocketCategoryStyle.Reusable);
-    perks.unshift(); // remove the archetype perk
-    let i = 0;
-    // Only pick 4 perks and purposefully ignore the origin perk
-    // on some guns as the link format doesn't support them.
-    for (const perk of _.take(perks, 4)) {
-      perkValues[i] = perk.plugged?.plugDef.hash ?? 0;
-      i++;
-    }
-    const masterwork = item.sockets.allSockets.find(isWeaponMasterworkSocket);
-    perkValues[4] = masterwork?.plugged?.plugDef.hash ?? 0;
-    const weaponMod = item.sockets.allSockets.find((s) =>
-      s.plugged?.plugDef.itemCategoryHashes?.includes(ItemCategoryHashes.WeaponModsDamage)
-    );
-    perkValues[5] = weaponMod?.plugged!.plugDef.hash ?? 0;
+function buildLightGGSockets(item: DimItem) {
+  const perkValues = getWeaponSocketInfo(item);
 
-    return perkValues.join(',');
+  if (perkValues) {
+    return perkValues.originTrait
+      ? `?p=${perkValues.traits.join(',')},${perkValues.originTrait},${perkValues.masterwork},${
+          perkValues.weaponMod
+        }`
+      : `?p=${perkValues.traits.join(',')},${perkValues.masterwork},${perkValues.weaponMod}`;
   }
 
   return '';
+}
+
+/**
+ * D2Gunsmith's socket format is: [...<first four perks, padded out if necessary, masterwork, weapon mod].join(',')
+ */
+function buildGunsmithSockets(item: DimItem) {
+  const perkValues = getWeaponSocketInfo(item);
+
+  if (perkValues) {
+    return `${perkValues.traits.join(',')},${perkValues.masterwork},${perkValues.weaponMod}`;
+  }
+
+  return '';
+}
+
+/**
+ * Gathers general socket information for link generation in D2Gunsmith and Light.gg.
+ */
+function getWeaponSocketInfo(
+  item: DimItem
+): null | { traits: number[]; originTrait: number; masterwork: number; weaponMod: number } {
+  if (item.sockets && item.bucket?.inWeapons) {
+    const traits: number[] = [0, 0, 0, 0];
+    const perks = getSocketsWithStyle(item.sockets, DestinySocketCategoryStyle.Reusable);
+    perks.unshift(); // remove the archetype perk
+    let i = 0;
+    // First 4 perks are on all weapons
+    for (const perk of _.take(perks, 4)) {
+      traits[i] = perk.plugged?.plugDef.hash ?? 0;
+      i++;
+    }
+
+    const origin = item.sockets.allSockets.find((s) =>
+      s.plugged?.plugDef.itemCategoryHashes?.includes(ItemCategoryHashes.WeaponModsOriginTraits)
+    );
+    const originTrait = origin?.plugged!.plugDef.hash ?? 0;
+
+    // TODO: Map enhanced intrinsic frames with their corresponding stat masterworks
+    const masterworkSocket = item.sockets.allSockets.find(isWeaponMasterworkSocket);
+    const masterwork =
+      masterworkSocket?.plugged?.plugDef.plug.plugCategoryHash ===
+      PlugCategoryHashes.CraftingPlugsFrameIdentifiers
+        ? 0
+        : masterworkSocket?.plugged?.plugDef.hash ?? 0;
+
+    const weaponModSocket = item.sockets.allSockets.find((s) =>
+      s.plugged?.plugDef.itemCategoryHashes?.includes(ItemCategoryHashes.WeaponModsDamage)
+    );
+    const weaponMod = weaponModSocket?.plugged!.plugDef.hash ?? 0;
+
+    return { traits, originTrait, masterwork, weaponMod };
+  }
+
+  return null;
 }